requirements.txt
gevent==1.2.2
bottle
firebase_admin
requests
PyYAML
PyNacl
